fix: Fix bugs in populating float values and in performing sanity check before loading the vector. (#38)

+++ velox/connectors/clp/search_lib/ir/ClpIrVectorLoader.cpp
@@ -85,7 +85,7 @@ void ClpIrVectorLoader::loadInternal(
         break;
       }
       case ColumnType::Float: {
-        auto floatVector = vector->asFlatVector<float>();
+        auto floatVector = vector->asFlatVector<double>();
         floatVector->set(
             vectorIndex,
             value->get_immutable_view<::clp::ffi::value_float_t>());
